move changeset walker
