+++ crates/trie/trie/Cargo.toml
@@ -13,11 +13,12 @@ workspace = true
 
 [dependencies]
 # reth
-reth-primitives.workspace = true
-reth-execution-errors.workspace = true
-reth-db.workspace = true
 reth-db-api.workspace = true
+reth-db.workspace = true
+reth-execution-errors.workspace = true
+reth-primitives.workspace = true
 reth-stages-types.workspace = true
+reth-storage-errors.workspace = true
 reth-trie-common.workspace = true
 
 revm.workspace = true
@@ -50,7 +51,6 @@ reth-chainspec.workspace = true
 reth-primitives = { workspace = true, features = ["test-utils", "arbitrary"] }
 reth-db = { workspace = true, features = ["test-utils"] }
 reth-provider = { workspace = true, features = ["test-utils"] }
-reth-storage-errors.workspace = true
 reth-trie-common = { workspace = true, features = ["test-utils", "arbitrary"] }
 
 # trie
